Restructured the code by createing components files
