+++ index.html
@@ -1,16 +1,14 @@
 <!DOCTYPE html>
 <html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./index.css" />
     <title>React-Tut</title>
-</head>
+  </head>
 
-<body>
+  <body>
     <div id="root"></div>
-    <script type="module" src="./App.js"></script>
-</body>
-
+    <script type="module" src="./src/App.js"></script>
+  </body>
 </html>
