+++ src/Components/RestaurentCard.js
@@ -1,0 +1,28 @@
+import { CDN_URL } from "../Utils/constants";
+
+const inlinneStyle = {
+  backgroundColor: "#f0f0f0",
+};
+
+const RestaurentCard = (props) => {
+  const { resData } = props;
+  const { cloudinaryImageId, name, cuisines, avgRating, costForTwo } =
+    resData?.info;
+  const { deliveryTime } = resData?.info?.sla;
+  return (
+    <div className="res-card" style={inlinneStyle}>
+      <img
+        className="res-logo"
+        alt="res-logo"
+        src={CDN_URL + cloudinaryImageId}
+      ></img>
+      <h3>{name}</h3>
+      <h4>{cuisines.join(", ")}</h4>
+      <h4>{avgRating}</h4>
+      <h4>{costForTwo}</h4>
+      <h4>{deliveryTime} minutes</h4>
+    </div>
+  );
+};
+
+export default RestaurentCard;
