+++ src/Utils/constants.js
@@ -1,0 +1,5 @@
+export const CDN_URL =
+  "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/";
+
+export const LOGO_URL =
+  "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR4BdwR2X16RrHC8QyJ5NXEZh83oY65HMqjBg&s";
